ticket:212 cache metadata when adding results to cart by clicking Add all to cart

+++ cghub/apps/core/tests/tests.py
@@ -37,22 +37,36 @@ class WithCacheTestCase(TestCase):
         # >>> get_cache_file_name('xml_text=6d5%2A', True)
         # u'/tmp/wsapi/427dcd2c78d4be27efe3d0cde008b1f9.xml'
 
+        if not self.cache_files:
+            return
+
         TEST_DATA_DIR = 'cghub/test_data/'
         if not os.path.exists(settings.WSAPI_CACHE_DIR):
             os.makedirs(settings.WSAPI_CACHE_DIR)
         for f in self.cache_files:
-            shutil.copy(
-                os.path.join(TEST_DATA_DIR, f),
-                os.path.join(settings.WSAPI_CACHE_DIR, f)
-            )
-        if self.cache_files:
-            self.default_results = objectify.fromstring(
-                open(os.path.join(settings.WSAPI_CACHE_DIR, self.cache_files[0])).read())
-            self.default_results_count = len(self.default_results.findall('Result'))
+            path_from = os.path.join(TEST_DATA_DIR, f)
+            if os.path.exists(path_from):
+                shutil.copy(
+                    path_from,
+                    os.path.join(settings.WSAPI_CACHE_DIR, f)
+                )
+        path = os.path.join(settings.WSAPI_CACHE_DIR, self.cache_files[0])
+        if not os.path.exists(path):
+            return
+        self.default_results = objectify.fromstring(
+            open(os.path.join(settings.WSAPI_CACHE_DIR, self.cache_files[0])).read())
+        self.default_results_count = len(self.default_results.findall('Result'))
 
     def tearDown(self):
         for f in self.cache_files:
-            os.remove(os.path.join(settings.WSAPI_CACHE_DIR, f))
+            path = os.path.join(settings.WSAPI_CACHE_DIR, f)
+            if not os.path.exists(path):
+                continue
+            if os.path.isdir(path):
+                # remove cart cache
+                shutil.rmtree(path)
+            else:
+                os.remove(path)
 
 
 class CoreTestCase(WithCacheTestCase):
